+++ kedro-gr3/conf/base/catalog.yml
@@ -1,0 +1,3 @@
+torillas:
+  type: pandas.CSVDataset
+  filepath: data/01_raw/tortilla_prices_raw.csv
