+++ kedro-gr3/pyproject.toml
@@ -1,0 +1,31 @@
+[build-system]
+requires = [ "setuptools",]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "kedro_gr3"
+readme = "README.md"
+dynamic = [ "dependencies", "version",]
+
+[project.scripts]
+kedro-gr3 = "kedro_gr3.__main__:main"
+
+[tool.kedro]
+package_name = "kedro_gr3"
+project_name = "kedro-gr3"
+kedro_init_version = "0.19.5"
+tools = [ "Kedro Viz",]
+example_pipeline = "False"
+source_dir = "src"
+
+[project.entry-points."kedro.hooks"]
+
+[tool.setuptools.dynamic.dependencies]
+file = "requirements.txt"
+
+[tool.setuptools.dynamic.version]
+attr = "kedro_gr3.__version__"
+
+[tool.setuptools.packages.find]
+where = [ "src",]
+namespaces = false
