+++ kedro-gr3/requirements.txt
@@ -1,0 +1,8 @@
+ipython>=8.10
+jupyterlab>=3.0
+kedro~=0.19.5
+kedro-datasets>=3.0; python_version >= "3.9"
+kedro-datasets>=1.0; python_version < "3.9"
+kedro-telemetry>=0.3.1
+kedro-viz>=6.7.0
+notebook
