+++ kedro-gr3/src/kedro_gr3/__init__.py
@@ -1,0 +1,4 @@
+"""kedro-gr3
+"""
+
+__version__ = "0.1"
