--- kedro/pyproject.toml
@@ -1,31 +1,0 @@
-[build-system]
-requires = [ "setuptools",]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "kedro"
-readme = "README.md"
-dynamic = [ "dependencies", "version",]
-
-[project.scripts]
-kedro = "kedro.__main__:main"
-
-[tool.kedro]
-package_name = "kedro"
-project_name = "kedro"
-kedro_init_version = "0.19.5"
-tools = [ "Kedro Viz",]
-example_pipeline = "False"
-source_dir = "src"
-
-[project.entry-points."kedro.hooks"]
-
-[tool.setuptools.dynamic.dependencies]
-file = "requirements.txt"
-
-[tool.setuptools.dynamic.version]
-attr = "kedro.__version__"
-
-[tool.setuptools.packages.find]
-where = [ "src",]
-namespaces = false
