--- kedro/src/kedro/__init__.py
@@ -1,4 +1,0 @@
-"""kedro
-"""
-
-__version__ = "0.1"
